move ProtocolCode class to own file
